chore(storage): Conditionally copying AWSAmplifyStressTests and AWSS3StoragePluginTests

+++ AmplifyPlugins/Storage/Tests/StorageHostApp/copy_configuration.sh
@@ -17,6 +17,16 @@ if [ ! -d "$SOURCE_DIR" ]; then
 fi
 
 mkdir -p "$DESTINATION_DIR"
-cp "$SOURCE_DIR/AWSS3StoragePluginTests-amplifyconfiguration.json" "$DESTINATION_DIR/amplifyconfiguration.json"
+
+if [ -f "$SOURCE_DIR/AWSAmplifyStressTests-amplifyconfiguration.json" ]; then
+    cp "$SOURCE_DIR/AWSAmplifyStressTests-amplifyconfiguration.json" "$DESTINATION_DIR/amplifyconfiguration.json"
+    exit 0
+fi
+
+if [ -f "$SOURCE_DIR/AWSS3StoragePluginTests-amplifyconfiguration.json" ]; then
+    cp "$SOURCE_DIR/AWSS3StoragePluginTests-amplifyconfiguration.json" "$DESTINATION_DIR/amplifyconfiguration.json"
+    exit 0
+fi
 
 exit 0
+
